simulator/requirements.txt
speechrecognition
pyttsx3
openai
supabase
python-dotenv
pyaudio
